feat: keep deprecated createRetryable on the package root
Re-add the auto-detecting `createRetryable` as a deprecated export from the
package root so existing `import { createRetryable } from 'ai-retry'` keeps
working. Its JSDoc points to the per-model entry points
(`ai-retry/<family>-model`), which are typed per family and resolve gateway
strings for every family.

The canonical dispatcher impl is restored to
`src/internal/create-retryable-model.ts`; `test-utils` re-exports it (instead
of holding its own copy) so the test helper and the public root export share
one implementation.
