+++ src/internal/test-utils.ts
@@ -17,7 +17,6 @@ import type {
   EmbeddingModel,
   EmbeddingModelCallOptions,
   EmbeddingModelEmbed,
-  GatewayLanguageModelId,
   ImageModel,
   ImageModelCallOptions,
   ImageModelGenerate,
@@ -27,16 +26,16 @@ import type {
   LanguageModelResult,
   LanguageModelStream,
   LanguageModelStreamPart,
-  RetryableModelOptions,
   RetryContext,
   RetryErrorAttempt,
   RetryResultAttempt,
 } from '../types.js';
-import { createRetryable as createRetryableEmbeddingModel } from '../embedding-model/create-retryable.js';
-import { createRetryable as createRetryableImageModel } from '../image-model/create-retryable.js';
-import { createRetryable as createRetryableLanguageModel } from '../language-model/create-retryable.js';
-import { isEmbeddingModel, isImageModel, isModel } from './guards.js';
-import { resolveLanguageModel } from './resolve-model.js';
+
+/**
+ * Re-exported for test convenience: the auto-detecting factory builds a
+ * retryable model for any family from a single import.
+ */
+export { createRetryable } from './create-retryable-model.js';
 
 type StreamText = Parameters<typeof streamText>[0];
 type GenerateText = Parameters<typeof generateText>[0];
@@ -161,60 +160,6 @@ export class MockImageModel implements ImageModel {
       return doGenerate;
     });
   }
-}
-
-/**
- * Test-only auto-detecting factory. Resolves the base model, infers its
- * family, and delegates to the matching family builder. Mirrors the way
- * the per-family entry points (`ai-retry/<family>-model`) are wired, so a
- * single import can build a retryable model for any family in tests. A
- * bare string is resolved as a language model.
- */
-export function createRetryable<MODEL extends LanguageModel>(
-  options: Omit<RetryableModelOptions<LanguageModel>, 'model'> & {
-    model: GatewayLanguageModelId;
-  },
-): LanguageModel;
-export function createRetryable<MODEL extends LanguageModel>(
-  options: RetryableModelOptions<MODEL>,
-): LanguageModel;
-export function createRetryable<MODEL extends EmbeddingModel>(
-  options: RetryableModelOptions<MODEL>,
-): EmbeddingModel;
-export function createRetryable<MODEL extends ImageModel>(
-  options: RetryableModelOptions<MODEL>,
-): ImageModel;
-export function createRetryable(
-  options:
-    | RetryableModelOptions<LanguageModel>
-    | RetryableModelOptions<EmbeddingModel>
-    | RetryableModelOptions<ImageModel>
-    | (Omit<RetryableModelOptions<LanguageModel>, 'model'> & {
-        model: GatewayLanguageModelId;
-      }),
-): LanguageModel | EmbeddingModel | ImageModel {
-  const model = isModel(options.model)
-    ? options.model
-    : resolveLanguageModel(options.model);
-
-  if (isEmbeddingModel(model)) {
-    return createRetryableEmbeddingModel({
-      ...options,
-      model,
-    } as RetryableModelOptions<EmbeddingModel>);
-  }
-
-  if (isImageModel(model)) {
-    return createRetryableImageModel({
-      ...options,
-      model,
-    } as RetryableModelOptions<ImageModel>);
-  }
-
-  return createRetryableLanguageModel({
-    ...options,
-    model,
-  } as RetryableModelOptions<LanguageModel>);
 }
 
 export const generateTextResult = (text: string): LanguageModelGenerate => ({
